service/requirements.txt
Flask==0.10.1
Flask-Cors==2.1.2
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.7
oauthlib==1.0.3
python-instagram==1.3.2
requests==2.8.1
requests-oauthlib==0.5.0
simplejson==3.8.1
six==1.10.0
tweepy==3.4.0
Werkzeug==0.11.2
